Some Refactor
Changed the name of some classes, for better legibility
