+++ dialogs/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
 
-from .user_profile_dialog import TradeDialog
+from .trade_dialog import TradeDialog
 
 __all__ = ["TradeDialog"]
